Move santizer classinto the integration test bundle
