+++ src/Pim/Bundle/ApiBundle/tests/integration/Controller/Media/ListMediaFileIntegration.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\ApiBundle\tests\integration\Controller\Media;
 
-use Akeneo\Test\Integration\MediaSanitizer;
+use Akeneo\Test\IntegrationTestsBundle\Sanitizer\MediaSanitizer;
 use Symfony\Component\HttpFoundation\Response;
 
 class ListMediaFileIntegration extends AbstractMediaFileTestCase
