+++ src/Pim/Bundle/VersioningBundle/tests/integration/Normalizer/Flat/ProductIntegration.php
@@ -3,7 +3,7 @@
 namespace tests\integration\Pim\Bundle\VersioningBundle\Normalizer\Flat;
 
 use Akeneo\Test\Integration\Configuration;
-use Akeneo\Test\Integration\MediaSanitizer;
+use Akeneo\Test\IntegrationTestsBundle\Sanitizer\MediaSanitizer;
 use Akeneo\Test\Integration\TestCase;
 
 /**
