+++ src/Pim/Component/Catalog/tests/integration/Updater/Setter/MediaAttributeSetterIntegration.php
@@ -3,7 +3,7 @@
 namespace Pim\Component\Catalog\tests\integration\Updater\Setter;
 
 use Akeneo\Test\Integration\Configuration;
-use Akeneo\Test\Integration\MediaSanitizer;
+use Akeneo\Test\IntegrationTestsBundle\Sanitizer\MediaSanitizer;
 use Akeneo\Test\Integration\TestCase;
 
 /**
